+++ catalog/app/components/Preview/load.js
@@ -1,17 +1,15 @@
 import * as React from 'react'
 
 import * as Audio from './loaders/Audio'
-import * as Csv from './loaders/Csv'
 import * as Echarts from './loaders/Echarts'
-import * as Excel from './loaders/Excel'
 import * as Fcs from './loaders/Fcs'
 import * as Html from './loaders/Html'
 import * as Image from './loaders/Image'
 import * as Json from './loaders/Json'
 import * as Markdown from './loaders/Markdown'
 import * as Notebook from './loaders/Notebook'
-import * as Parquet from './loaders/Parquet'
 import * as Pdf from './loaders/Pdf'
+import * as Tabular from './loaders/Tabular'
 import * as Text from './loaders/Text'
 import * as Vcf from './loaders/Vcf'
 import * as Video from './loaders/Video'
@@ -19,21 +17,19 @@ import * as Voila from './loaders/Voila'
 import * as fallback from './loaders/fallback'
 
 const loaderChain = [
-  Csv,
-  Excel,
   Fcs,
   Echarts, // should be before Json, or TODO: add "type is not 'echarts'" to Json.detect
   Json,
   Markdown,
   Voila, // should be before Notebook, or TODO: add "type is not 'voila'" to Notebook.detect
   Notebook,
-  Parquet,
   Pdf,
   Vcf,
   Html,
   Image,
   Video,
   Audio,
+  Tabular,
   Text,
   fallback,
 ]